Adapt to only select hearthstones from a pledged covenant
In 9.0.5, Blizzard changed the covenant hearthstone toys to require that
you be pledged to that covenant.  The Necrolord one was apparently always
configured this way, so the addon was incorrectly selecting it for other
covenant members who had it in their toybox, producing a "you must be
pledged to the Necrolords" error, but now this is happening for all
four of the covanent toys.

Ref: https://www.wowhead.com/news/covenant-hearthstone-toys-can-only-be-used-by-players-of-that-covenant-321415